code/requirements.txt
django 
djangorestframework
pyyaml
requests
django-cors-headers
Celery
django-crispy-forms